requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pathlib>=1.0.1
feedparser>=6.0.10
pandas>=2.0.0
pyyaml>=6.0.1
lxml>=4.9.3
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0
questionary>=2.0.1
rich>=13.7.0
windows-curses>=2.3.1; platform_system == "Windows"
Jinja2>=3.1.2
openai>=1.3.0
aiohttp>=3.9.0
tenacity>=8.2.3
tiktoken>=0.5.1
pyarrow>=14.0.1
jsonschema>=4.20.0
transformers>=4.36.0
datasets>=2.15.0
torch>=2.1.0
tqdm>=4.66.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
wandb>=0.16.0
tensorboard>=2.15.0 
xmltodict
zipfile
io
tqdm
fastapi 
"uvicorn[standard]"